+++ src/components/primitives/number/types.ts
@@ -11,4 +11,11 @@ export type NumberProps = BaseNullableInputElementProps<number>
          * @default 1
         */
         step?: number;
+
+        /**
+         * Reset value to `defaultValue` (or `0` if not specified) on focus out when the field is empty.
+         * Prevents the input from having no value.
+         * @default false
+         */
+        resetToDefaultOnBlur?: boolean;
     };